Revert "some root src folder cleanup"
This reverts commit e40a0f6f39cfbb6738219fafb3a3030a12e35888.
