+++ bsb/connectivity/detailed/shared.py
@@ -2,7 +2,7 @@ from itertools import chain
 from functools import reduce, cache
 import numpy as np
 from ...storage import Chunk
-from ...core.reporting import warn
+from ...reporting import warn
 from ...exceptions import *
