+++ bsb/connectivity/detailed/touch_detection.py
@@ -1,7 +1,7 @@
 import numpy as np
 from ..strategy import ConnectionStrategy
 from .shared import Intersectional
-from ...core.reporting import report, warn
+from ...reporting import report, warn
 from random import sample as sample_elements
 from ... import config
 from ...config import types
