+++ bsb/placement/satellite.py
@@ -1,7 +1,7 @@
 from .strategy import PlacementStrategy
 import math, numpy as np
 from ..exceptions import *
-from ..core.reporting import report, warn
+from ..reporting import report, warn
 from .. import config
 from ..config import types, refs
 from itertools import chain
