+++ bsb/simulators/arbor/devices/probe.py
@@ -7,7 +7,7 @@ from ....simulation.results import (
 )
 from ....simulation.device import Patternless
 from ....exceptions import *
-from ....core.reporting import report, warn
+from ....reporting import report, warn
 import numpy as np
