+++ bsb/topology/region.py
@@ -6,7 +6,7 @@ from ._layout import Layout, RhomboidData
 from .. import config
 from ..config import types, refs
 from ..exceptions import *
-from ..core.reporting import warn
+from ..reporting import warn
 import numpy as np
 import abc
